modal/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.8.5
    # via
    #   langchain
    #   modal
    #   openai
aiosignal==1.3.1
    # via aiohttp
aiostream==0.4.5
    # via modal
anyio==3.7.1
    # via
    #   fastapi
    #   httpcore
    #   starlette
    #   watchfiles
asgiref==3.7.2
    # via modal
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   sigtools
beautifulsoup4==4.12.2
    # via
    #   -r requirements.in
    #   googlesearch-python
    #   markdownify
    #   unstructured
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   modal
    #   requests
chardet==5.2.0
    # via unstructured
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
clean-text[gpl]==0.6.0
    # via -r requirements.in
click==8.1.7
    # via
    #   modal
    #   nltk
    #   typer
    #   uvicorn
cloudpickle==2.2.1
    # via modal
customerio==2.1
    # via -r requirements.in
dataclasses-json==0.5.14
    # via
    #   langchain
    #   unstructured
emoji==1.7.0
    # via
    #   clean-text
    #   unstructured
fastapi==0.103.1
    # via
    #   -r requirements.in
    #   modal
filetype==1.2.0
    # via unstructured
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
ftfy==6.1.1
    # via clean-text
gender-guesser==0.4.0
    # via -r requirements.in
googlesearch-python==1.2.3
    # via -r requirements.in
greenlet==2.0.2
    # via playwright
grpclib==0.4.3
    # via modal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h2==4.1.0
    # via grpclib
hpack==4.0.0
    # via h2
html2text==2020.1.16
    # via -r requirements.in
httpcore==0.17.3
    # via httpx
httptools==0.6.0
    # via uvicorn
httpx==0.24.1
    # via python-telegram-bot
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via modal
joblib==1.3.2
    # via nltk
langchain==0.0.293
    # via -r requirements.in
langdetect==1.0.9
    # via -r requirements.in
langsmith==0.0.38
    # via langchain
lxml==4.9.3
    # via unstructured
markdown-it-py==3.0.0
    # via rich
markdownify==0.11.6
    # via -r requirements.in
marshmallow==3.20.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
modal==0.52.3461
    # via modal-client
modal-client==0.52.3461
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   grpclib
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nltk==3.8.1
    # via unstructured
numexpr==2.8.6
    # via langchain
numpy==1.26.0
    # via
    #   langchain
    #   numexpr
    #   pandas
openai==0.28.0
    # via -r requirements.in
packaging==23.1
    # via marshmallow
pandas==2.1.0
    # via -r requirements.in
pendulum==2.1.2
    # via -r requirements.in
playwright==1.38.0
    # via
    #   -r requirements.in
    #   playwright-stealth
playwright-stealth==1.0.6
    # via -r requirements.in
protobuf==4.24.3
    # via modal
psycopg2==2.9.7
    # via -r requirements.in
pydantic==1.10.12
    # via
    #   fastapi
    #   langchain
    #   langsmith
    #   sqlmodel
pyee==9.0.4
    # via playwright
pygments==2.16.1
    # via rich
python-dateutil==2.8.2
    # via
    #   pandas
    #   pendulum
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
python-magic==0.4.27
    # via unstructured
python-telegram-bot==20.5
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pytzdata==2020.1
    # via pendulum
pyyaml==6.0.1
    # via
    #   langchain
    #   uvicorn
regex==2023.8.8
    # via
    #   nltk
    #   tiktoken
requests==2.31.0
    # via
    #   customerio
    #   googlesearch-python
    #   langchain
    #   langsmith
    #   openai
    #   tiktoken
    #   unstructured
rich==13.5.3
    # via modal
sigtools==4.0.1
    # via synchronicity
six==1.16.0
    # via
    #   langdetect
    #   markdownify
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==1.4.41
    # via
    #   langchain
    #   sqlmodel
sqlalchemy2-stubs==0.0.2a35
    # via sqlmodel
sqlmodel==0.0.8
    # via -r requirements.in
starlette==0.27.0
    # via fastapi
synchronicity==0.5.3
    # via modal
tabulate==0.9.0
    # via unstructured
tblib==2.0.0
    # via modal
tenacity==8.2.3
    # via langchain
tiktoken==0.5.1
    # via -r requirements.in
toml==0.10.2
    # via modal
tqdm==4.66.1
    # via
    #   nltk
    #   openai
typer==0.9.0
    # via modal
types-certifi==2021.10.8.3
    # via modal
types-toml==0.10.8.7
    # via modal
typing-extensions==4.8.0
    # via
    #   fastapi
    #   modal
    #   pydantic
    #   pyee
    #   sqlalchemy2-stubs
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.3
    # via pandas
unidecode==1.3.6
    # via clean-text
unstructured==0.10.15
    # via -r requirements.in
urllib3==2.0.4
    # via requests
uvicorn[standard]==0.23.2
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.20.0
    # via
    #   modal
    #   uvicorn
wcwidth==0.2.6
    # via ftfy
websockets==11.0.3
    # via uvicorn
yarl==1.9.2
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata
